Cleaned a bit more the MapsFragment
