+++ app/src/main/java/com/github/sdpteam15/polyevents/model/map/RouteMapHelper.kt
@@ -16,6 +16,7 @@ import com.github.sdpteam15.polyevents.model.map.LatLngOperator.scalar
 import com.github.sdpteam15.polyevents.model.map.LatLngOperator.time
 import com.github.sdpteam15.polyevents.model.observable.Observable
 import com.github.sdpteam15.polyevents.model.observable.ObservableList
+import com.github.sdpteam15.polyevents.view.activity.admin.ZoneManagementActivity
 import com.github.sdpteam15.polyevents.view.fragments.MapsFragment
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.Marker
@@ -627,4 +628,6 @@ object RouteMapHelper {
             tempLatLng[1] = endMarker!!.position
         }
     }
+
+
 }
